I have already done the methods of the interface, but I don't implement the interface beacuse when I compile the programs appears a mistake that say that the classes where I'm implementing the interface, have to be abstract.

+++ src/model/Price.java
@@ -1,5 +1,5 @@
 package model;
 public interface Price{
 	public double calculateMarketPrice();
-	public int calculateStarsLevel();
+	public double calculateStarsLevel();
 }
